+++ spec/blather/stanza/x_spec.rb
@@ -64,6 +64,13 @@ describe Blather::Stanza::X do
     expect(n.title).to eq("goodbye")
   end
 
+  it 'escapes title properly' do
+    n = Blather::Stanza::X.new :form
+    expect(n.title).to eq(nil)
+    n.title = "<a&a>"
+    expect(n.title).to eq("<a&a>")
+  end
+
   it 'has an instructions attribute' do
     n = Blather::Stanza::X.new :form
     expect(n.instructions).to eq(nil)
@@ -71,6 +78,13 @@ describe Blather::Stanza::X do
     expect(n.instructions).to eq("Please fill in this form")
     n.instructions = "goodbye"
     expect(n.instructions).to eq("goodbye")
+  end
+
+  it 'escapes instructions properly' do
+    n = Blather::Stanza::X.new :form
+    expect(n.instructions).to eq(nil)
+    n.instructions = "<a&a>"
+    expect(n.instructions).to eq("<a&a>")
   end
 
   it 'inherits a list of fields' do
@@ -197,6 +211,13 @@ describe Blather::Stanza::X::Field do
     expect(n.value).to eq('book1')
     n.value = 'book2'
     expect(n.value).to eq('book2')
+  end
+
+  it 'allows multiple values' do
+    n = Blather::Stanza::X::Field.new 'subject', 'list-multi', 'Music from the time of Shakespeare'
+    expect(n.value).to eq(nil)
+    n.value = ['book<&1>', 'book2']
+    expect(n.value).to eq(['book<&1>', 'book2'])
   end
 
   it 'allows setting options' do